+++ src/language-css/index.ts
@@ -1,0 +1,3 @@
+import { Reprinter, ReprinterOptions } from "./reprinter";
+
+export { Reprinter as CssReprinter, ReprinterOptions as CssReprinterOptions };