+++ src/language-js/sortJsxElement/index.ts
@@ -1,12 +1,17 @@
 import { Comment } from "estree";
 
-import {
-  getContextGroups,
-  getSpreadGroups,
-  reorderValues
-} from "../utilities/sort-utils";
+import { getContextGroups, reorderValues } from "../../utilities/sort-utils";
+import { getSpreadGroups } from "../utilities/sort-utils";
 
 export interface SortJsxElementOptions {}
+
+interface JsxAttribute {
+  range?: [number, number];
+  type: string;
+  name: {
+    name: string;
+  };
+}
 
 export function sortJsxElement(
   jsxElement: any,
@@ -16,17 +21,17 @@ export function sortJsxElement(
 ) {
   let newFileContents = fileContents.slice();
 
-  let allNodes = jsxElement.openingElement.attributes;
+  let allNodes: JsxAttribute[] = jsxElement.openingElement.attributes;
 
   // Any time there is a spread operator, we need to sort around it... moving it could cause functionality changes
-  let spreadGroups: any[] = getSpreadGroups(allNodes);
+  let spreadGroups = getSpreadGroups(allNodes);
 
   for (let nodes of spreadGroups) {
     let groupings = getContextGroups(nodes, comments, fileContents);
 
     groupings.forEach(element => {
-      let unsorted: any[] = element.nodes;
-      let sorted: any[] = element.nodes.slice().sort((a, b) => {
+      let unsorted = element.nodes;
+      let sorted = element.nodes.slice().sort((a, b) => {
         return a.name.name.localeCompare(b.name.name);
       });