+++ changelog.md
@@ -2,6 +2,7 @@
 
 - Added `css`, `scss` and `less` file support
 - Upgraded typescript-estree@7.0.0
+- Fixed dependent static property sort for javascript's sortClassContents
 
 ### 2.3.2